requirements.txt
argon2-cffi==20.1.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
langid==1.1.6
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.6
notebook==6.1.3
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySide2==5.15.0
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
Send2Trash==1.5.0
shiboken2==5.15.0
six==1.15.0
socketio==0.2.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.1.0
